refactor(tests): split flows.rs by feature domain
